U-H-V#1
Useless Harmless Virus #1 is a simple script that when opened on victimes computer  kills the explorer.exe process and puts a copy of himself in the windows startup folder so that whenever the pc is started the script will end explorer task. To remove the "virus" from the victime's computer either use the script given in the repo or press Win+R and type "shell:startup" delete the file named windows10.bat and the script will be succesfully removed from the computer. For any problems please write in the issues tab i will try my best to respond at every message.